Fix bug when insert consensus from subread is gap-only
Resolves issue #17

The problem occurs with inserts reported on telomeric repeat regions on
the reference. When insert + flanking are extracted from subreads and
consensus is taken, the consensus may end up being shorter than the
original flanking regions without insert from the reference. This
results in nonsensical insert consensus being reported, e.g. all-gaps.

Another issue is that SPOA v4.0.3 reports trailing gap-only columns with
pairwise alignments under some conditions, which should be removed, and
which may lead to the wrong coordinates being chosen for the insert.

To resolve both issues, (1) check that the insert + flanking is longer
than the original reference flanking regions, and skip if this is not
true, and (2) when aligning insert + flanking consensus to reference,
remove gap-only columns.